Sweet Victory (NorWestIN)
The winner takes all
It's the thrill of one more kill
The last one to fall
Will never sacrifice their will
Don't ever look back
On the wind closing in
The only attack
Were their wings on the wind
Oh the daydream begins
And it's sweet, sweet, sweet victory, yeah!
And it's ours for the taking
It's ours for the fight
In the sweet, sweet, sweet victory, yeah!
And the world is last to fall